feat(dashboard): refactor dashboard (4)
move code into `src` subdirectory
