+++ dashboard/src/components/layout/footer.tsx
