+++ dashboard/src/pages/sensor/[sensorName].tsx
@@ -1,11 +1,11 @@
-import { SENSOR_IDS } from "@/utils/constants";
-import { ICONS } from "@/components/icons";
+import { SENSOR_IDS } from "@/src/utils/constants";
+import { ICONS } from "@/src/components/icons";
 import Link from "next/link";
 import { useState } from "react";
-import { useNetworkStore } from "@/utils/state";
-import { determinSensorStatus, renderTimeString } from "@/utils/functions";
+import { useNetworkStore } from "@/src/utils/state";
+import { determinSensorStatus, renderTimeString } from "@/src/utils/functions";
 import { maxBy } from "lodash";
-import { VARIANT_TO_PILL_COLOR } from "@/utils/colors";
+import { VARIANT_TO_PILL_COLOR } from "@/src/utils/colors";
 
 export function getStaticPaths() {
   return {
